+++ examples/drawing_samples/draw_samples.py
@@ -18,8 +18,7 @@ from crystal_diffusion.models.position_diffusion_lightning_model import \
     PositionDiffusionLightningModel
 from crystal_diffusion.oracle.energies import compute_oracle_energies
 from crystal_diffusion.samplers.variance_sampler import NoiseParameters
-from crystal_diffusion.samples_and_metrics.sampling import \
-    create_batch_of_samples
+from crystal_diffusion.samples.sampling import create_batch_of_samples
 from crystal_diffusion.utils.logging_utils import setup_analysis_logger
 
 logger = logging.getLogger(__name__)
