+++ tests/samples_and_metrics/test_sampling.py
@@ -6,8 +6,7 @@ from crystal_diffusion.generators.position_generator import (
     PositionGenerator, SamplingParameters)
 from crystal_diffusion.namespace import (CARTESIAN_POSITIONS,
                                          RELATIVE_COORDINATES, UNIT_CELL)
-from crystal_diffusion.samples_and_metrics.sampling import \
-    create_batch_of_samples
+from crystal_diffusion.samples.sampling import create_batch_of_samples
 from crystal_diffusion.utils.basis_transformations import \
     get_positions_from_coordinates
